General Prediction and Coefficient Extractors
Generalized pred.func() assignment; there are only 3 varieties of prediction functions, which are now defined at the top, as opposed to separately within each fit function definition. The appropriate prediction function is assigned depending on the method it serves.

Similarly, a coef.func() function was added, which defines how the coefficients are to be extracted, when applicable. Additionally, this function joins the return list along with the pred.func and model elements.

This step builds towards making the coefficients available to the fit.subgroup() function, and in turn, validate.subgroup().