Update fit_losses to account for matching.id
Logic added in to have clusters belong to the same folds whenever LASSO is used so that every matched group member always belongs to the same fold. If user specifies both a matching.id and a foldid, the user-specified foldid is currently ignored.

Still need to consider if/how the matching ID should play a role in the cross-validation piece of the GBM models.